requirements.txt
asgiref==3.3.4
astroid==2.7.3
attrs==20.3.0
Authlib==0.15.4
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
awscli==1.20.36
backports.entry-points-selectable==1.1.0
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
beautifulsoup4==4.9.3
botocore==1.21.36
braceexpand==0.1.7
Brotli==1.0.9
cachelib==0.3.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
check50==3.3.3
click==7.1.2
colorama==0.4.3
compare50==1.2.4
constantly==15.1.0
cryptography==3.4.8
cs50==7.0.2
cssselect2==0.4.1
cycler==0.10.0
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.2
Django==3.2
django-admin==2.0.1
django-allauth==0.45.0
django-configuration==0.0.3
django-crispy-forms==1.12.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-polymorphic==3.0.0
django-six==1.0.4
djangorestframework==3.12.4
docutils==0.15.2
EditorConfig==0.12.3
filelock==3.0.12
Flask==1.1.4
Flask-Session==0.4.0
fonttools==4.26.2
greenlet==1.1.1
help50==3.0.5
html5lib==1.1
hyperlink==21.0.0
icdiff==2.0.4
idna==3.2
ikp3db==1.4.5
incremental==21.3.0
Inflector==2.0.12
intervaltree==2.1.0
isort==5.9.3
itsdangerous==1.1.0
jellyfish==0.8.8
Jinja2==2.11.3
jmespath==0.10.0
jsbeautifier==1.14.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lib50==3.0.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
natsort==7.1.1
nose==1.3.7
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.2
pexpect==4.8.0
Pillow==8.3.2
platformdirs==2.3.0
plotly==5.3.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydyf==0.1.1
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.10.2
pylint-django==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
python-dateutil==2.8.2
python-decouple==3.4
python-magic==0.4.24
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
render50==5.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3cmd==2.1.0
s3transfer==0.5.0
screen==1.0.1
service-identity==21.1.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
sqlparse==0.4.1
style50==2.7.5
submit50==3.1.0
tenacity==8.0.1
termcolor==1.1.0
tinycss2==1.1.0
toml==0.10.2
tqdm==4.62.2
Twisted==21.7.0
twython==3.9.1
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.7.2
weasyprint==53.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==3.0.1
xlwt==1.3.0
zope.interface==5.4.0
zopfli==0.1.8
